feat(sfMap): float map + doc-06 snowplough dust channel
The grid goes rgba16float, linear values end to end: pack-scale factors
deleted, /255 decodes deleted, CPU readback via the existing f16ToFloat
(GalaxySfMap.data is now Float32Array). oldActivity keeps its [0,1]
clamp — a real EMA bound, not a format artifact.

Snowplough per doc-06: gas stays a regenerating fuel channel; the swept
conserved quantity is a NEW dust channel in the z-slot freed by
deriving refractory from age. Ignition keeps dustFloorFraction (new
knob, default 0.2) of the cell's dust and deposits the rest on Moore
neighbors; colliding fronts SUM — the overshoot above ambient is the
rim, ceilinged at 8x only as a runaway guard. Dust CONSUMERS
(sfMapDustDensity et al.) still read gas x oldActivity; re-keying them
to the swept channel is the follow-up that makes walls visible.

+++ src/@types/galaxy/GalaxySfMap.ts
@@ -1,17 +1,19 @@
 /**
  * GalaxySfMap — one CPU-side readback of the SSPSF automaton's packed output
- * (`sfMapPack.wesl`): a log-polar RGBA8 grid, R=gas fraction, G=recent SF,
- * B=older activity (A unused). `data` is TIGHTLY packed (az*4 bytes per row,
- * no GPU `copyTextureToBuffer` row-alignment padding) — see
- * the galaxy tool's `createSfMapReadbacks` for where the 256-byte stride gets
- * stripped back out. `rMin`/`rMax` are the log-radial bounds THIS
- * readback's grid was built over (`sfMapGridRadius`), needed to invert
- * `sfMapRingRadius` when sampling.
+ * (`sfMapPack.wesl`): a log-polar RGBA grid, already decoded from the
+ * texture's rgba16float storage to LINEAR floats — R=gas fraction, G=recent
+ * SF, B=older activity, A=dust (conserved via the snowplough rule; A may
+ * exceed 1.0, see `sfMapPack.wesl`'s header). `data` is TIGHTLY packed (4
+ * floats per texel, row-major, no GPU `copyTextureToBuffer` row-alignment
+ * padding or f16 bit pattern) — see the galaxy tool's `createSfMapReadbacks`
+ * for where both get stripped/decoded. `rMin`/`rMax` are the log-radial
+ * bounds THIS readback's grid was built over (`sfMapGridRadius`), needed to
+ * invert `sfMapRingRadius` when sampling.
  */
 export type GalaxySfMap = {
   readonly az: number;
   readonly rings: number;
   readonly rMin: number;
   readonly rMax: number;
-  readonly data: Uint8Array;
+  readonly data: Float32Array;
 };
